Create GraphQL schema and generate interfaces
Created the GraphQL schema to be the Spartify API. This allows for development
to continue in parallel for the backend and frontend. Note that the backend
SQL table database models as specified during our Week 2 lab section should be
consistent with implementing this API.

Ran "npm run gen" to validate the schema and generate the TypeScript interfaces
for the GraphQL types.
https://github.com/rothfels/bespin/wiki/How-To-Guide

This is a breaking change, as the rest of the project still implements the class
website project.